organize/rename and use `TagBadge`
