--- ui-v2/src/components/task-runs/index.tsx
@@ -1,1 +1,0 @@
-export { TaskRuns } from "./task-run-details";
