+++ ui-v2/src/components/task-runs/task-run-details/index.ts
@@ -1,0 +1,1 @@
+export { TaskRunDetails } from "./task-run-details";
