+++ ui-v2/src/components/task-runs/task-run-details/task-run-details.test.tsx
@@ -1,19 +1,19 @@
 import { createFakeTaskRun } from "@/mocks";
 import { render } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import { TaskRuns } from ".";
+import { TaskRunDetails } from "./task-run-details";
 
-describe("TaskRuns", () => {
+describe("TaskRunDetails", () => {
 	it("should display flow run link with task name prefix", () => {
 		const taskRun = createFakeTaskRun({ name: "test-task-name" });
-		const screen = render(<TaskRuns taskRun={taskRun} />);
+		const screen = render(<TaskRunDetails taskRun={taskRun} />);
 
 		expect(screen.getByText("test")).toBeInTheDocument();
 	});
 
 	it("should display task run ID", () => {
 		const taskRun = createFakeTaskRun({ id: "test-task-id" });
-		const screen = render(<TaskRuns taskRun={taskRun} />);
+		const screen = render(<TaskRunDetails taskRun={taskRun} />);
 
 		expect(screen.getByText("test-task-id")).toBeInTheDocument();
 	});
@@ -23,7 +23,7 @@ describe("TaskRuns", () => {
 		const taskRun = createFakeTaskRun({
 			start_time: startTime,
 		});
-		const screen = render(<TaskRuns taskRun={taskRun} />);
+		const screen = render(<TaskRunDetails taskRun={taskRun} />);
 
 		expect(screen.getByText("Start Time").nextSibling).toBeInTheDocument();
 	});
@@ -32,7 +32,7 @@ describe("TaskRuns", () => {
 		const taskRun = createFakeTaskRun({
 			tags: ["tag1", "tag2", "tag3"],
 		});
-		const screen = render(<TaskRuns taskRun={taskRun} />);
+		const screen = render(<TaskRunDetails taskRun={taskRun} />);
 
 		expect(screen.getByText("tag1")).toBeInTheDocument();
 		expect(screen.getByText("tag2")).toBeInTheDocument();
@@ -40,7 +40,7 @@ describe("TaskRuns", () => {
 	});
 
 	it("should display empty state when no taskRun is provided", () => {
-		const screen = render(<TaskRuns taskRun={null} />);
+		const screen = render(<TaskRunDetails taskRun={null} />);
 
 		expect(
 			screen.getByText("No task run details available"),
@@ -54,7 +54,7 @@ describe("TaskRuns", () => {
 		const taskRun = createFakeTaskRun({
 			task_inputs: taskInputs,
 		});
-		const screen = render(<TaskRuns taskRun={taskRun} />);
+		const screen = render(<TaskRunDetails taskRun={taskRun} />);
 
 		expect(screen.getByText("Task Inputs")).toBeInTheDocument();
 		expect(screen.getByText(/test-param/)).toBeInTheDocument();
@@ -62,7 +62,7 @@ describe("TaskRuns", () => {
 
 	it("should display task configuration section", () => {
 		const taskRun = createFakeTaskRun();
-		const screen = render(<TaskRuns taskRun={taskRun} />);
+		const screen = render(<TaskRunDetails taskRun={taskRun} />);
 
 		expect(screen.getByText("Task configuration")).toBeInTheDocument();
 		expect(screen.getByText("Version")).toBeInTheDocument();
